git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@15875 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/e/ntl-interface.cpp
@@ -31,16 +31,15 @@ NTL::mat_ZZ *mutableMatrix_to_NTL_mat_ZZ(const MutableMatrix *M)
 MutableMatrix *mutableMatrix_from_NTL_mat_ZZ(const NTL::mat_ZZ *A)
 {
   // AGAIN: form the TRANSPOSE of A
-  int ncols = static_cast<int>(A->NumRows());
-  int nrows = static_cast<int>(A->NumCols());
-
+  size_t ncols = A->NumRows();
+  size_t nrows = A->NumCols();
   MutableMatrix *B = MutableMatrix::zero_matrix(globalZZ, nrows, ncols, false);
 
   mpz_t a;
   mpz_init(a);
 
-  for (int i=0; i<ncols; i++)
-    for (int j=0; j<nrows; j++)
+  for (size_t i=0; i<ncols; i++)
+    for (size_t j=0; j<nrows; j++)
       {
         if ((*A)(i+1,j+1) != 0)
           {
